% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_market_share_columns.R
\name{add_market_share_columns}
\alias{add_market_share_columns}
\title{Add market share columns to a scenario dataset}
\usage{
add_market_share_columns(data, reference_year)
}
\arguments{
\item{data}{A scenario dataset, like FIXME: Define an exported demo scenario.}

\item{reference_year}{The baseline year, against which the technology- and
sector- market shares will be calculated. Note: At the start year, tmsr = 1
and smsp =0 respectively.}
}
\value{
A scenario dataset, with the new columns \code{tmsr} and \code{smsp}.
}
\description{
Calculates and adds market share values (ie. technology market-share ratio
and sector market-share percentage) to a scenario dataset. A reference
start-year must be provided.
}
