% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p4b.R
\name{format_p4b}
\alias{format_p4b}
\title{Format scenario data for P4B}
\usage{
format_p4b(data)
}
\arguments{
\item{data}{A scenario dataset, like \link[pacta.scenario.preparation:weo_2021]{pacta.scenario.preparation::weo_2021}.}
}
\value{
A scenario dataset, with columns renamed to be consistent with
r2dii.analysis target_market_share input requirements.
}
\description{
Format scenario data for P4B
}
\examples{
library(dplyr)
weo_2022 \%>\%
  interpolate_yearly(source, scenario, sector, technology, scenario_geography, indicator, units) \%>\%
  filter(year >= 2022) \%>\%
  add_market_share_columns(reference_year = 2022) \%>\%
  format_p4b()
}
