% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p4b_ei.R
\name{format_p4b_ei}
\alias{format_p4b_ei}
\title{Format scenario data for P4B}
\usage{
format_p4b_ei(data)
}
\arguments{
\item{data}{A scenario dataset, like \link[pacta.scenario.preparation:weo_2021]{pacta.scenario.preparation::weo_2021}.}
}
\value{
A scenario dataset, with columns renamed to be consistent with
r2dii.analysis target_sda input requirements.
}
\description{
Format scenario data for P4B
}
