% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p4i.R
\name{format_p4i}
\alias{format_p4i}
\title{Format scenario data for P4I}
\usage{
format_p4i(data, green_techs)
}
\arguments{
\item{data}{A scenario dataset.}

\item{green_techs}{A list of green technologies. For these, a \code{direction} of
"increasing" will be assigned, and the \code{smsp} column will be used to assign
a \code{fair_share_perc}. Otherwise the \code{direction} will be \code{decreasing} and the
\code{tmsr} column will be used.}
}
\value{
A scenario dataset, with columns renamed to be consistent with
pacta.data.preparation input requirements.
}
\description{
Format scenario data for P4I
}
