% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_2023_technology_bridge.R
\docType{data}
\name{geco_2023_technology_bridge}
\alias{geco_2023_technology_bridge}
\title{An technology bridge for GECO 2023}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 152 rows and 2 columns.
}
\usage{
geco_2023_technology_bridge
}
\description{
Fake data about physical assets (e.g. wind turbine power plant
capacities), aggregated to company-level. These data are used to assess the
climate alignment of financial portfolios. It imitates data from
market-intelligence databases.
}
\examples{
head(geco_2023_technology_bridge)
}
\keyword{datasets}
