% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_yearly.R
\name{interpolate_yearly}
\alias{interpolate_yearly}
\title{Interpolate values in a dataset, by year.}
\usage{
interpolate_yearly(data, ...)
}
\arguments{
\item{data}{An input dataset. Must contain the columns \code{year} and \code{value}.}

\item{...}{Other grouping variables. \code{value} will be interpolated for each
group.}
}
\value{
A dataset with the column \code{value} interpolated linearly against the
column \code{year}.
}
\description{
Interpolate values in a dataset, by year.
}
