% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_geco_2022_scenario.R
\name{prepare_geco_2022_scenario}
\alias{prepare_geco_2022_scenario}
\title{Prepare GECO 2022 scenario data}
\usage{
prepare_geco_2022_scenario(
  geco_2022_automotive_raw,
  geco_2022_aviation_raw,
  geco_2022_fossil_fuels_15c_raw,
  geco_2022_fossil_fuels_ndc_raw,
  geco_2022_fossil_fuels_ref_raw,
  geco_2022_power_15c_raw,
  geco_2022_power_ndc_raw,
  geco_2022_power_ref_raw,
  geco_2022_steel_raw
)
}
\arguments{
\item{geco_2022_automotive_raw}{A raw GECO 2022 automotive scenario
data-frame.}

\item{geco_2022_aviation_raw}{A raw GECO 2022 aviation scenario data-frame.}

\item{geco_2022_fossil_fuels_15c_raw}{A raw GECO 2022 fossil fuels 1.5C
scenario data-frame.}

\item{geco_2022_fossil_fuels_ndc_raw}{A raw GECO 2022 fossil fuels NDC
scenario data-frame.}

\item{geco_2022_fossil_fuels_ref_raw}{A raw GECO 2022 fossil fuels reference
scenario data-frame.}

\item{geco_2022_power_15c_raw}{A raw GECO 2022 power 1.5C scenario
data-frame.}

\item{geco_2022_power_ndc_raw}{A raw GECO 2022 power NDC scenario data-frame.}

\item{geco_2022_power_ref_raw}{A raw GECO 2022 power reference scenario
data-frame.}

\item{geco_2022_steel_raw}{A raw GECO 2022 steel scenario data-frame.}
}
\value{
A prepared GECO 2022 scenario data-frame.
}
\description{
Prepare GECO 2022 scenario data
}
