% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_geco_2023_scenario.R
\name{prepare_geco_2023_scenario}
\alias{prepare_geco_2023_scenario}
\title{Prepare GECO 2023 scenario data}
\usage{
prepare_geco_2023_scenario(
  technology_bridge,
  geco_2023_aviation_15c_raw,
  geco_2023_aviation_ndc_raw,
  geco_2023_aviation_ref_raw,
  geco_2023_fossil_fuels_15c_raw,
  geco_2023_fossil_fuels_ndc_raw,
  geco_2023_fossil_fuels_ref_raw,
  geco_2023_power_cap_15c_raw,
  geco_2023_power_cap_ndc_raw,
  geco_2023_power_cap_ref_raw,
  geco_2023_steel_15c_raw,
  geco_2023_steel_ndc_raw,
  geco_2023_steel_ref_raw,
  geco_2023_supplement_15c_raw,
  geco_2023_supplement_ndc_raw,
  geco_2023_supplement_ref_raw
)
}
\arguments{
\item{technology_bridge}{A technology bridge data-frame.}

\item{geco_2023_aviation_15c_raw}{A raw GECO 2023 automotive 1.5 scenario data-frame.}

\item{geco_2023_aviation_ndc_raw}{A raw GECO 2023 automotive NDC scenario data-frame.}

\item{geco_2023_aviation_ref_raw}{A raw GECO 2023 automotive Reference scenario data-frame.}

\item{geco_2023_fossil_fuels_15c_raw}{A raw GECO 2023 fossil fuels 1.5 scenario data-frame.}

\item{geco_2023_fossil_fuels_ndc_raw}{A raw GECO 2023 fossil fuels NDC scenario data-frame.}

\item{geco_2023_fossil_fuels_ref_raw}{A raw GECO 2023 fossil fuels Reference scenario data-frame.}

\item{geco_2023_power_cap_15c_raw}{A raw GECO 2023 power capacity 1.5 scenario data-frame.}

\item{geco_2023_power_cap_ndc_raw}{A raw GECO 2023 power capacity NDC scenario data-frame.}

\item{geco_2023_power_cap_ref_raw}{A raw GECO 2023 power capacity Reference scenario data-frame.}

\item{geco_2023_steel_15c_raw}{A raw GECO 2023 steel 1.5 scenario data-frame.}

\item{geco_2023_steel_ndc_raw}{A raw GECO 2023 steel NDC scenario data-frame.}

\item{geco_2023_steel_ref_raw}{A raw GECO 2023 steel Reference scenario data-frame.}

\item{geco_2023_supplement_15c_raw}{A raw GECO 2023 supplemental 1.5 scenario data-frame.}

\item{geco_2023_supplement_ndc_raw}{A raw GECO 2023 supplemental NDC scenario data-frame.}

\item{geco_2023_supplement_ref_raw}{A raw GECO 2023 supplemental Reference scenario data-frame.}
}
\value{
A prepared GECO 2023 scenario data-frame.
}
\description{
Prepare GECO 2023 scenario data
}
