% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_isf_2021_scenario.R
\name{prepare_isf_2021_scenario}
\alias{prepare_isf_2021_scenario}
\title{Prepare ISF 2021 scenario data}
\usage{
prepare_isf_2021_scenario(isf_2021_power_raw, isf_2021_not_power_raw)
}
\arguments{
\item{isf_2021_power_raw}{A tidyxl data frame containing a raw import of
\code{NZAOA_raw_data_power.xlsx}.}

\item{isf_2021_not_power_raw}{A tidyxl data frame containing a raw import of
\code{NZAOA_rawdata_notpower_P4I.xlsx}.}
}
\value{
A prepared ISF 2021 scenario data-frame.
}
\description{
Prepare ISF 2021 scenario data
}
