% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_isf_2023_scenario.R
\name{prepare_isf_2023_scenario}
\alias{prepare_isf_2023_scenario}
\title{Prepare ISF 2023 scenario data}
\usage{
prepare_isf_2023_scenario(
  isf_2023_scope_global_raw,
  isf_2023_annex_countries_raw
)
}
\arguments{
\item{isf_2023_scope_global_raw}{A tidyxl data frame (with a \code{formats}
attribute) with a raw ISF Scope Global 2023 import.}

\item{isf_2023_annex_countries_raw}{A list of tidyxl data frames (with a \code{formats} attribute) containing the raw import of each of the Annex Countries xlsx files for ISF 2023.}
}
\value{
A prepared ISF 2023 scenario data-frame.
}
\description{
Prepare ISF 2023 scenario data
}
