% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_weo_2024_hybrid_in_ev_scenario.R
\name{prepare_weo_2024_hybrid_in_ev_scenario}
\alias{prepare_weo_2024_hybrid_in_ev_scenario}
\title{Prepare WEO 2024 scenario data}
\usage{
prepare_weo_2024_hybrid_in_ev_scenario(
  weo_2024_ext_data_regions_raw,
  weo_2024_ext_data_world_raw,
  weo_2024_fig_chptr_3_raw,
  iea_global_ev_2024_raw,
  iea_sales_share_ev,
  mpp_ats_raw,
  config_name
)
}
\arguments{
\item{weo_2024_ext_data_regions_raw}{A data frame containing a raw
\code{WEO2024_Extended_Data_Regions.csv} import.}

\item{weo_2024_ext_data_world_raw}{A data frame containing a raw
\code{WEO2024_Extended_Data_World.csv} import.}

\item{weo_2024_fig_chptr_3_raw}{A tidyxl data frame containing a raw import
of \code{WEO2024_Figures_Chapter_03.xlsx}.}

\item{iea_global_ev_2024_raw}{A data frame containing a raw import of 'electric-vehicle-sales-by-region-and-scenario-2030-and-2035.xlsx'
sheet : electric-vehicle-sales-by-regio}

\item{iea_sales_share_ev}{A data frame containing a raw import of 'electric-vehicle-sales-by-region-and-scenario-2030-and-2035.xlsx'
\itemize{
\item sheet: electric vehicle share-ev
}}

\item{mpp_ats_raw}{A tidyxl data frame containing a raw import of `2022-08-12
\itemize{
\item MPP ATS - RPK and GHG intensity.xlsx`.
}}

\item{config_name}{A string providing the timestamp used to prepare scenario}
}
\value{
A prepared WEO 2024 scenario data-frame.
}
\description{
Prepare WEO 2024 scenario data
}
