% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_regions.R
\docType{data}
\name{scenario_regions}
\alias{scenario_regions}
\title{A dataset of countries contained in different scenario regions.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1492 rows and 4 columns.
}
\usage{
scenario_regions
}
\description{
This dataset contains a map of each scenario region name, along with all
countries defined within that region (as per the country's ISO2 code).
}
\examples{
head(scenario_regions)
}
\keyword{datasets}
