% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p4b.R
\name{format_p4b}
\alias{format_p4b}
\title{Format scenario data for P4B}
\usage{
format_p4b(data)
}
\arguments{
\item{data}{A scenario dataset.}
}
\value{
A scenario dataset, with columns renamed to be consistent with
r2dii.analysis target_market_share input requirements.
}
\description{
Format scenario data for P4B
}
